pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.5",
    "langchain-huggingface>=0.3.0",
    "langchain-tavily>=0.2.7",
    "langchainhub>=0.1.21",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.46.1",
    "tavily-python>=0.7.9",
    "tiktoken>=0.9.0",
]
